scrapers/requirements.txt
scrapling>=0.4.0
curl_cffi>=0.15.0
playwright>=1.50.0
patchright>=1.59.0
pyyaml>=6.0
